fixed prototype function type mismatch for the 32 bit change

+++ pulsetrain.h
@@ -113,13 +113,13 @@ uint8_t pSetupTimers();
 uint8_t pNewPTrain();
 uint8_t pAttach(uint8_t ptrain_index, int pin, timers16bit_t timer);
 
-uint8_t pSetPulseUS(uint8_t ptrain_index, uint16_t period, 
-                        uint16_t pulse_width, uint16_t period_num_limit);
-uint8_t _pSetPulseUS(uint8_t ptrain_index, uint16_t period, 
-                        uint16_t pulse_width,  uint16_t period_num_limit, 
+uint8_t pSetPulseUS(uint8_t ptrain_index, uint32_t period, 
+                        uint32_t pulse_width, uint16_t period_num_limit);
+uint8_t _pSetPulseUS(uint8_t ptrain_index, uint32_t period, 
+                        uint32_t pulse_width,  uint16_t period_num_limit, 
                         uint16_t prescale);
-uint8_t pSetPulseOnlyUS(uint8_t ptrain_index, uint16_t pulse_width);
-uint8_t pSetPeriodOnlyUS(uint8_t ptrain_index, uint16_t period);
+uint8_t pSetPulseOnlyUS(uint8_t ptrain_index, uint32_t pulse_width);
+uint8_t pSetPeriodOnlyUS(uint8_t ptrain_index, uint32_t period);
 uint8_t pSetPeriodNumberOnly(uint8_t ptrain_index, uint16_t period_num_limit);
 uint16_t pGetPulseCounts(uint8_t ptrain_index);
 uint16_t pGetPeriodCounts(uint8_t ptrain_index);
